% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_vl_ascending_distance}
\alias{get_vl_ascending_distance}
\title{Get voice-leading ascending distance}
\usage{
get_vl_ascending_distance(e1, e2, elt_type)
}
\description{
Returns the ascending distance from <e1> to <e2>,
after Tymoczko (2006): the smallest nonnegative real
number x such that, if p is a pitch with label e1,
then p + x has label e2. <elt-type> may be either
"pitch" or "pc; it denotes whether <e1>
and <e2> are pitches or pitch classes.
Vectorised over <e2>.
}
