% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_vl_elt_distance}
\alias{get_vl_elt_distance}
\title{Get voice-leading element distance}
\usage{
get_vl_elt_distance(e1, e2, elt_type)
}
\description{
Computes the voice-leading distance between two
elements <e1> and <e2> given <elt_type>. <elt_type>
can be "pc", in which case the distance
is computed between two pitch classes, or
"pitch", in which cases distance is computed
between two pitches.
}
