% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_vl_set_distance}
\alias{get_vl_set_distance}
\title{Get voice-leading set distance}
\usage{
get_vl_set_distance(s1, s2, elt_type, norm)
}
\description{
Computes the voice-leading distance between two
ordered sets of the same length, <s1> and <s2>,
given <elt-type> (which can either be
:pitch or :pitch-class; see vl-elt-distance for
definition) and <norm>, which determines the
norm used: currently <norm> can take values
of :euclidean, :taxicab, and :infinity.
}
