% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_vl}
\alias{min_vl}
\title{Get minimal voice-leading}
\usage{
min_vl(s1, s2, elt_type = "pc", norm = "taxicab",
  preserve_bass = FALSE)
}
\arguments{
\item{s1}{The first set to be compared; numeric vector,
with each number corresponding to either a pitch or a pitch class.
Duplicates are permitted, and they will be retained. Order does not matter.}

\item{s2}{The second set to be compared; see \code{s1}.}

\item{elt_type}{Can be either \code{pitch} or \code{pc};
determines whether \code{s1} and \code{s2} are interpreted as
pitches or pitch classes.}

\item{norm}{Can be either \code{euclidean}, \code{taxicab}, or \code{infinity}.
Each of these identify different norms.}

\item{preserve_bass}{Logical scalar;
if TRUE, the first element of each set is considered to be the bass note,
and the voice leading is constrained to preserve these bass notes.
Only relevant if \code{elt_type == "pc"}.}
}
\value{
A list of three values:
1) the minimal voice-leading distance;
2) the corresponding start voicing;
3) the corresponding end voicing.
}
\description{
Computes the minimal voice-leading between two sets of pitches or pitch classes, using the polynomial-time algorithm described in Tymoczko (2006).
}
\references{
\insertRef{Tymoczko2006}{minVL}
}
