% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_vls}
\alias{min_vls}
\title{Minimum voice leadings}
\usage{
min_vls(s1_list, s2, elt_type = "pc", norm = "taxicab",
  preserve_bass = FALSE)
}
\arguments{
\item{s1_list}{A list of voice leadings to be analysed.
Each element should be a numeric vector,
with each number corresponding to either a pitch or a pitch class.
Duplicates are permitted, and they will be retained. Order does not matter.}

\item{s2}{A voice leading to be analysed, represented as a numeric vector.}

\item{elt_type}{Can be either \code{pitch} or \code{pc};
determines whether \code{s1} and \code{s2} are interpreted as
pitches or pitch classes.}

\item{norm}{Can be either \code{euclidean}, \code{taxicab}, or \code{infinity}.
Each of these identify different norms.}

\item{preserve_bass}{Logical scalar;
if TRUE, the first element of each set is considered to be the bass note,
and the voice leading is constrained to preserve these bass notes.
Only relevant if \code{elt_type == "pc"}.}
}
\value{
An integer vector of the same length as \code{s1_list},
where element i corresponds to the minimal voice leading between
the ith element of \code{s1_list} and \code{s2}.
}
\description{
This is a vectorised version of \code{\link{min_vl}}.
}
